+++ CHANGELOG.md
@@ -1,4 +1,15 @@
+## 2.0.1
+
+#### 6-Jul-2022
+
+- Added Symfony 5.4+ service as example
+- Updated README.md with short docs about class usage and symfony integration
+
 ## 2.0.0
+
+#### 5-Jul-2022
+
+- First relase as composer package
 
 #### 11-Jun-2022